chore: address comments 3
🔒 Scanned for secrets using gitleaks 8.28.0
